% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CryptProtectData}
\alias{CryptProtectData}
\title{Wrapper for the Windows DPAPI function CryptProtectData()}
\usage{
CryptProtectData(pwd, salt)
}
\arguments{
\item{pwd}{The raw representation of the string top be encrypted}

\item{salt}{A vector of randomly generated bytes}
}
\description{
This function is intended to be an internal helper file for the package.
Both parameters are currently necessary, and bust be between 1 and 256 bytes
in length. The salt should be generated using PKI::PKI.random(), or some
other cryptographically stron random number generator.
}
