% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LockerAccessUtils.R
\name{PurgeExpiredPasswords}
\alias{PurgeExpiredPasswords}
\title{Remove all passwords in the table that are beyond their expiration date}
\usage{
PurgeExpiredPasswords(df)
}
\description{
Remove all passwords in the table that are beyond their expiration date
}
