% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LockerAccessUtils.R
\name{ReadPasswordTable}
\alias{ReadPasswordTable}
\title{Retrieve the password table, purging expired passwords along the way}
\usage{
ReadPasswordTable()
}
\description{
Retrieve the password table, purging expired passwords along the way
}
