% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RememberPassword.R
\name{RememberPassword}
\alias{RememberPassword}
\title{Prompt the user for a password and store it for later use}
\usage{
RememberPassword(user, application, expiration.date = NA,
  expiration.duration = NA)
}
\arguments{
\item{user}{The username associated with the password}

\item{expiration.date}{Coercible to a datetime object of the form "y-m-d h:m:s"}

\item{expiration.duration}{A names list indicating the number of days, hours, and minutes until the password should expire}

\item{applicaiton}{The application that the username and password should be used for}
}
\description{
Store the user's password on disk in an encrypted form. This encrypted
form can only be decrypted when the user is logged into their Windows
account. Other users cannot decrypt it even if they have access to the hard
drive with administrator privileges.
}
\details{
Encryption is done using 3DES with a 128 bit salt to protect password privacy
across multiple systems even if they use the same password.
}
