% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LockerAccessUtils.R
\name{RetrieveEncryptedPassword}
\alias{RetrieveEncryptedPassword}
\title{Retrieve an encrypted password from the data frame}
\usage{
RetrieveEncryptedPassword(username, application)
}
\description{
Retrieve an encrypted password from the data frame
}
