% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrievePassword.R
\name{RetrievePassword}
\alias{RetrievePassword}
\title{Retrieve a stored password from the locker}
\usage{
RetrievePassword(user, application = NA)
}
\arguments{
\item{user}{A string indicating the user being logged in}

\item{application}{A string indicating the application being logged into}
}
\description{
This function returns the plaintext password that a user previously stored.
The password is returned as a single string so that it can be simply passed
inline to whatever application requires it.
}
