% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LockerAccessUtils.R
\name{StoreEncryptedPassword}
\alias{StoreEncryptedPassword}
\title{Add an encrypted password to the data frame}
\usage{
StoreEncryptedPassword(username, application, expiration, salt, password)
}
\description{
Add an encrypted password to the data frame
}
